+++ server/types/game.go
@@ -42,8 +42,8 @@ type Game struct {
 	End    time.Time `json:"end"`
 	Status string    `json:"status"`
 
-	P1Checkers [25]int `json:"p1checkers"` // arr[0] is bar
-	P2Checkers [25]int `json:"p2checkers"` // arr[0] is bar
+	P1Checkers [25]int8 `json:"p1checkers"` // arr[0] is bar
+	P2Checkers [25]int8 `json:"p2checkers"` // arr[0] is bar
 
 	Double      uint64 `json:"double"`
 	DoubleOwner string `json:"double-owner"`
